+++ ext/openssl/config0.m4
@@ -26,7 +26,7 @@ PHP_ARG_WITH([openssl-argon2],
 
 if test "$PHP_OPENSSL" != "no"; then
   PHP_NEW_EXTENSION([openssl],
-    [openssl.c openssl_pwhash.c xp_ssl.c],
+    [openssl.c openssl_pwhash.c openssl_backend_common.c openssl_backend_v1.c openssl_backend_v3.c xp_ssl.c],
     [$ext_shared])
   PHP_SUBST([OPENSSL_SHARED_LIBADD])
   PHP_SETUP_OPENSSL([OPENSSL_SHARED_LIBADD],
